third_party/blink/web_tests/fast/forms/select/customizable-select/disallowed-select-descendants-console-message.html

<!DOCTYPE html>
<link rel=author href="mailto:ansollan@microsoft.com">
<link rel=help href="https://issues.chromium.org/issues/347890366">

<style>
  select, ::picker(select) {
    appearance: base-select;
  }
</style>

<!-- Should print one console message. -->
<select>
  <label>label</label>
</select>

<!-- Should print two console messages. -->
<select>
  <input>
  <input>
</select>

<!-- Should print one console messages. -->
<select>
  <option>
    <input>
  </option>
</select>

<!-- Should print one console messages. -->
<select>
  <option>
    <datalist>
    </datalist>
  </option>
</select>

<!-- Should print one console messages. -->
<select>
  <option>
    <object>
    </object>
  </option>
</select>

<!-- Phrasing content but not Interactive content allowed. -->
<select>
  <option>
    <textarea></textarea>
  </option>
</select>

<!-- <button> (if present) should be the first child of <select>. -->
<select>
  <option></option>
  <option></option>
  <button>button</button>
</select>

<!-- <legend> (if present) should be the first child of <optgroup>. -->
<select>
  <optgroup>
    <option>..</option>
    <legend>legend</legend>
  </optgroup>
</select>

<select id="delayed" style="display:none">
  <input>
  <option></option>
</select>

<script>
  requestAnimationFrame(() => {
    document.getElementById('delayed').style.display = '';
    requestAnimationFrame(() => {
        testRunner.notifyDone();
    });
  });
</script>

<script>
testRunner.waitUntilDone();
testRunner.dumpAsText();
</script>
